// automation/packages/server/sandbox/src/lib/types.ts
export type SandboxSettings = {
    EXECUTION_MODE: string
    DEV_PIECES: string[]
    ENVIRONMENT: string
    REUSE_SANDBOX: string | undefined
    FLOW_TIMEOUT_SECONDS: number
    MAX_FILE_SIZE_MB: number
    MAX_FLOW_RUN_LOG_SIZE_MB: number
    NETWORK_MODE: NetworkMode
    SANDBOX_MEMORY_LIMIT: string
    SANDBOX_PROPAGATED_ENV_VARS: string[]
    SSRF_ALLOW_LIST: string[]
    ENFORCE_CONNECTION_PIECE_BINDING: boolean
    WORKER_GROUP_ID?: string | undefined
    PROJECT_WORKER?: boolean | undefined
}